Finish repainting the user interface.
This turned out to be a major refactoring and improves the plugin
across the board. The editor user interface now is way more clear
about what it is aiming at and its interactivity got improved to
support the operator while tuning its knobs.

Along the lines, appropriate exception and error handling has been
introduced to signal error conditions and warnings to the user which 
helps tremendously when setting up and operating this plugin.