Add support for Graylog Geolocation plugin
Creates a new Location Data option ("string"), which makes it possible
to query coordinates that are stored as a string in a single field.

The field should be formatted as "latitude,longitude".

Graylog's Geolocation plugin stores data this way instead of as a
geohash. This has made it difficult for Graylog users to use this plugin.